requirements.txt
sentence-transformers==2.2.2
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
ctransformers==0.2.5
pinecone-client==2.2.4
langchain==0.0.225
-e .




